[Vulnerabilities: A2] - Update -> apostrophe-caches-redis Security update has been released for redis to fix the vulnerability.
When a client is in monitoring mode, the regex begin used to detected monitor messages could cause exponential backtracking on some strings. This issue could lead to a denial of service.

We only have to update to the latest 3.x series. In Version 4, they changed how you connect to redis and it's complicated.

Action: apostrophe-caches-redis@2.1.4 using redis@2.8.0 which has vulnerability but fixed in redis@3.1.1. So apos team needs to update redis in apostrophe-caches-redis

This is an exponential backtracking risk if keys are constructed in a certain way. Apostrophe is not constructing keys in this way, but redis should still be updated.